+++ src/Controller/Back/PaintingController.php
@@ -4,10 +4,12 @@ namespace App\Controller\Back;
 
 use App\Entity\CompanyPainting;
 use App\Entity\Painting;
+use App\Entity\Preparation;
 use App\Form\PaintingType;
 use App\Repository\CompanyPaintingRepository;
 use App\Repository\PaintingRepository;
 use App\Repository\UserRepository;
+use App\Security\Voter\GeneralVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -15,7 +17,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route("/company/services/paintings")]
+#[Route("/morgue/services/peintures")]
 #[IsGranted("ROLE_COMPANY")]
 class PaintingController extends AbstractController
 {
@@ -25,29 +27,70 @@ class PaintingController extends AbstractController
         $user = $userRep->find($this->getUser());
         $company = $user->getCompany();
 
+        $paintings = $paintingRep->findBy([
+            'deletedAt' => null,
+        ], [
+            'name' => 'ASC'
+        ]);
         $paintingsSuscribedByCompany = $paintingRep->getAllByCompany($company);
 
+        foreach ($paintings as $painting) {
+            $painting->canBeDeleted = $this->canBeDeleted($painting);
+            $painting->canBeSwitched = $this->canBeSwitched($painting);
+
+        }
+
+        foreach ($paintingsSuscribedByCompany as $painting) {
+            $painting->canBeSwitched = $this->canBeSwitched($painting);
+        }
+
+
         return $this->render("back/company/services/paintings/index.html.twig", [
-            "paintings" => $paintingRep->findAll(),
+            "paintings" => $paintings,
             "paintingsSuscribedByCompany" => $paintingsSuscribedByCompany
         ]);
     }
 
 
     /* THANATOS DATABASE RELATED */
+    private function canBeDeleted(Painting $painting): bool
+    {
+        $em = $this->getDoctrine()->getManager();
+        $preparations = $em->getRepository(Preparation::class)->findAll();
+
+        if (!empty($preparations)) $preparations = array_map(function ($p) {
+            if ($p->getModelMaterial() !== null)
+                return $p->getModelMaterial()->getId();
+        }, $preparations);
+
+        return empty($painting->getCompanyPaintings()->toArray()) && !in_array($painting->getId(), $preparations);
+    }
+
+    private function canBeSwitched(Painting $painting): bool
+    {
+        $em = $this->getDoctrine()->getManager();
+        $preparations = $em->getRepository(Preparation::class)->findAll();
+
+        if (!empty($preparations)) $preparations = array_map(function ($p) {
+            if ($p->getModelMaterial() !== null)
+                return $p->getModelMaterial()->getId();
+        }, $preparations);
+
+        return !in_array($painting->getId(), $preparations);
+    }
 
 
     #[Route('/details/{id}', name: 'details_painting')]
-    public function details_painting(PaintingRepository $paintingRep, int $id): Response
+    public function details_painting(Painting $painting): Response
     {
-       /* dump($id);
-        dd($paintingRep->find($id));*/
+        $this->denyAccessUnlessGranted(GeneralVoter::VIEW_EDIT, $painting);
+
         return $this->render("back/company/services/paintings/details.html.twig", [
-            "painting" => $paintingRep->find($id)
+            "painting" => $painting
         ]);
     }
 
-    #[Route('/create', name: 'create_painting')]
+    #[Route('/créer', name: 'create_painting')]
     public function create_painting(Request $request, EntityManagerInterface $em, PaintingRepository $paintingRep): Response
     {
         $painting = new Painting();
@@ -65,10 +108,12 @@ class PaintingController extends AbstractController
         ]);
     }
 
-    #[Route('/modify/{id}', name: 'modify_painting')]
-    public function modify_paintings(Request $request, EntityManagerInterface $em, PaintingRepository $paintingRep, int $id): Response
+    #[Route('/modifier/{id}', name: 'modify_painting')]
+    public function modify_paintings(Request $request, EntityManagerInterface $em, PaintingRepository $paintingRep, Painting $painting): Response
     {
-        $painting = $paintingRep->find($id);
+
+        $this->denyAccessUnlessGranted(GeneralVoter::VIEW_EDIT, $painting);
+
         $form = $this->createForm(PaintingType::class, $painting);
         $form->handleRequest($request);
 
@@ -83,10 +128,12 @@ class PaintingController extends AbstractController
         ]);
     }
 
-    #[Route('/delete/{id}', name: 'delete_painting')]
-    public function delete_painting(EntityManagerInterface $em, PaintingRepository $paintingRep, int $id): Response
+    #[Route('/supprimer/{id}', name: 'delete_painting')]
+    public function delete_painting(EntityManagerInterface $em, PaintingRepository $paintingRep, Painting $painting): Response
     {
-        $painting = $paintingRep->find($id);
+        $this->denyAccessUnlessGranted(GeneralVoter::VIEW_EDIT, $painting);
+        if (!$this->canBeDeleted($painting)) dd('page error');
+
         $media = $painting->getMedia();
         $em->remove($media);
         $em->remove($painting);
@@ -100,23 +147,28 @@ class PaintingController extends AbstractController
     /* EXPOSED TO THE CLIENTS */
 
     #[Route('/switch/{id}', name: 'switch_painting')]
-    public function switch_painting(int $id, EntityManagerInterface $em, UserRepository $userRep, CompanyPaintingRepository $companyPaintingRep, PaintingRepository $paintingRep) : Response {
+    public function switch_painting(Painting $painting, EntityManagerInterface $em, UserRepository $userRep, CompanyPaintingRepository $companyPaintingRep, PaintingRepository $paintingRep): Response
+    {
         $user = $userRep->find($this->getUser());
         $company = $user->getCompany();
-        $painting = $paintingRep->find($id);
 
         // TODO : Meilleur vérification plus tard (genre theme désactivé par thanatos)
         if (empty($company)) {
             $this->addFlash("error", "Une erreur est survenue. Veuillez réessayez ultérieurement");
             return $this->redirectToRoute("view_paintings");
         }
+        $this->denyAccessUnlessGranted(GeneralVoter::VIEW_EDIT, $painting);
 
         $companyPainting = $companyPaintingRep->getOneByCompanyAndPainting($company, $painting);
 
         if ($companyPainting) {
+            if (!$this->canBeSwitched($painting)) {
+                $this->addFlash("error", "La couleur " . $painting->getName() . " ne peut être retiré car il est utilisé dans une commande");
+                return $this->redirectToRoute("view_paintings");
+            }
             $em->remove($companyPainting);
             $em->flush();
-            $this->addFlash("success", "La couleur ".$painting->getName()." ne sera plus disponible pour les clients");
+            $this->addFlash("success", "La couleur " . $painting->getName() . " ne sera plus disponible pour les clients");
             return $this->redirectToRoute("view_paintings");
         }
 
@@ -127,7 +179,7 @@ class PaintingController extends AbstractController
         $em->persist($companyPainting);
         $em->flush();
 
-        $this->addFlash("success", "La couleur ".$painting->getName()." est désormais disponibles pour les clients");
+        $this->addFlash("success", "La couleur " . $painting->getName() . " est désormais disponibles pour les clients");
         return $this->redirectToRoute("view_paintings");
 
     }
