+++ src/Controller/Back/ThemeController.php
@@ -3,32 +3,42 @@
 namespace App\Controller\Back;
 
 use App\Entity\CompanyTheme;
+use App\Entity\Preparation;
+use App\Entity\Theme;
 use App\Repository\CompanyRepository;
 use App\Repository\CompanyThemeRepository;
 use App\Repository\ThemeRepository;
 use App\Repository\UserRepository;
+use App\Security\Voter\GeneralVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route("/company/services/themes")]
+#[Route("/morgue/services/themes")]
 #[IsGranted("ROLE_COMPANY")]
 class ThemeController extends AbstractController
 {
     #[Route('/', name: 'view_company_themes')]
-    public function view_company_themes(ThemeRepository $themeRep, CompanyThemeRepository $companyThemeRep, UserRepository $userRep ) : Response {
+    public function view_company_themes(ThemeRepository $themeRep, CompanyThemeRepository $companyThemeRep, UserRepository $userRep): Response
+    {
         $user = $userRep->find($this->getUser());
         $company = $user->getCompany();
 
-        $themes = $themeRep->findAll();
+        $themes = $themeRep->findBy([
+            'deletedAt' => null,
+        ], [
+            'type' => 'ASC',
+            'name' => 'ASC'
+        ]);
         $themesSuscribedByCompany = $themeRep->getAllByCompany($company);
 
         $detailedThemes = [];
 
         foreach ($themes as $theme) {
             $detailedThemes[] = [
+                "canBeSwitched" => $this->canBeSwitched($theme),
                 "theme" => $theme,
                 "isSuscribed" => in_array($theme, $themesSuscribedByCompany, true),
             ];
@@ -39,11 +49,25 @@ class ThemeController extends AbstractController
         ]);
     }
 
+    private function canBeSwitched(Theme $theme): bool
+    {
+
+        $company = $this->getUser()->getCompany();
+
+        $em = $this->getDoctrine()->getManager();
+        $preparations = $em->getRepository(Preparation::class)->getPreparationsByCompany($company);
+
+        return empty($preparations);
+    }
+
     #[Route('/switch/{id}', name: 'switch_company_themes')]
-    public function switch_company_themes(EntityManagerInterface $em,  ThemeRepository $themeRep, int $id, CompanyRepository $companyRep, UserRepository $userRep, CompanyThemeRepository $companyThemeRep) : Response {
+    public function switch_company_themes(EntityManagerInterface $em, ThemeRepository $themeRep, Theme $theme, CompanyRepository $companyRep, UserRepository $userRep, CompanyThemeRepository $companyThemeRep): Response
+    {
+
         $user = $userRep->find($this->getUser());
         $company = $user->getCompany();
-        $theme = $themeRep->find($id);
+
+        $this->denyAccessUnlessGranted(GeneralVoter::VIEW_EDIT, $theme);
 
         // TODO : Meilleur vérification plus tard (genre theme désactivé par thanatos)
         if (!empty($company)) {
@@ -51,9 +75,15 @@ class ThemeController extends AbstractController
             $companyTheme = $companyThemeRep->getOneByCompanyAndTheme($company, $theme);
 
             if ($companyTheme) {
+
+                if (!$this->canBeSwitched($theme)) {
+                    $this->addFlash("error", "Le thème " . $theme->getName() . " ne peut être retiré car il est utilisé dans une commande");
+                    return $this->redirectToRoute("view_company_themes");
+                }
+
                 $em->remove($companyTheme);
                 $em->flush();
-                $this->addFlash("success", "Vous vous êtes désinscrit du thème ".$theme->getName().".");
+                $this->addFlash("success", "Vous vous êtes désinscrit du thème " . $theme->getName() . ".");
                 return $this->redirectToRoute("view_company_themes");
             }
 
@@ -63,7 +93,7 @@ class ThemeController extends AbstractController
 
             $em->persist($companyTheme);
             $em->flush();
-            $this->addFlash("success", "Vous vous êtes inscrit au thème ".$theme->getName().".");
+            $this->addFlash("success", "Vous vous êtes inscrit au thème " . $theme->getName() . ".");
             return $this->redirectToRoute("view_company_themes");
         }
